Defer wallet cancellation and remove extra comments
We previously anticipated needing to persist wallets in case the process
dies and we need to recover. In this case, if we don't persist wallets
on creation, we will have to create new wallets on restart which is
wasteful. However, this is actually not required, we are ok with this
waste, so embrace the simplification it provides.